requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
braceexpand==0.1.7
cachetools==5.5.2
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.1
deepspeed==0.16.3
einops==0.8.1
filelock==3.16.1
fire==0.7.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.2.3
future==1.0.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
guided-filter-pytorch==3.7.5
hjson==3.1.0
huggingface-hub==0.29.3
imageio==2.35.1
imgaug==0.4.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
jax==0.4.13
jinja2==3.1.6
kiwisolver==1.4.7
kornia==0.7.3
kornia-rs==0.1.10
lazy-loader==0.4
lightning-utilities==0.3.0
lit==15.0.7
lpips==0.1.4
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
natsort==8.4.0
networkx==3.1
ninja==1.11.1.3
nvidia-ml-py==12.570.86
nvitop==1.4.2
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.30.0
opencv-python==4.11.0.86
opt-einsum==3.4.0
packaging==24.2
pandas==2.0.3
propcache==0.2.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.1
pytorch-msssim==1.0.0
pytz==2025.1
pywavelets==1.4.1
pyyaml==6.0.2
regex==2024.11.6
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
tifffile==2023.7.10
timm==0.6.13
tokenizers==0.13.3
torchmetrics==0.11.4
tqdm==4.67.1
transformers==4.33.3
triton==2.0.0
typing-extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
webdataset==0.2.100
werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
